+++ api/src/main/java/org/flab/api/domain/event/domain/musical/Cast.java
@@ -11,7 +11,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
-import org.flab.api.domain.event.dto.event.response.CastResponse;
+import org.flab.api.domain.event.dto.event.response.musical.CastResponse;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
