--- api/src/main/java/org/flab/api/domain/event/repository/CharacterRepository.java
@@ -1,10 +1,0 @@
-package org.flab.api.domain.event.repository;
-
-import org.flab.api.domain.event.domain.musical.Character;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface CharacterRepository extends JpaRepository<Character, Long> {
-    List<Character> findByEventId(Long id);
-}
