+++ api/src/main/java/org/flab/api/domain/event/repository/concert/ConcertRepository.java
@@ -1,4 +1,4 @@
-package org.flab.api.domain.event.repository;
+package org.flab.api.domain.event.repository.concert;
 
 import org.flab.api.domain.event.domain.concert.Concert;
 import org.springframework.data.jpa.repository.JpaRepository;
