+++ api/src/main/java/org/flab/api/domain/event/repository/concert/ConcertRepositoryImpl.java
@@ -1,4 +1,4 @@
-package org.flab.api.domain.event.repository;
+package org.flab.api.domain.event.repository.concert;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
