+++ api/src/test/java/org/flab/api/service/ConcertServiceTest.java
@@ -2,7 +2,7 @@ package org.flab.api.service;
 
 import org.flab.api.domain.event.domain.EventType;
 import org.flab.api.domain.event.domain.concert.Concert;
-import org.flab.api.domain.event.repository.ConcertRepository;
+import org.flab.api.domain.event.repository.concert.ConcertRepository;
 import org.flab.api.domain.event.service.ConcertService;
 import org.flab.api.global.exception.CustomException;
 import org.flab.api.global.exception.ErrorCode;
